# contracts/openapi/une-platform-api-v1.yaml
openapi: 3.1.0
info:
  title: UNE 재난안전 AI 문서 통합플랫폼 API
  version: 1.0.0
  description: ADR v1.1 및 API·DB·화면별 Sequence 상세설계서 기준 121개 내부 API 계약 초안
# NOTE(CC-003/ADR-20): GET /health is a confirmed out-of-contract operational endpoint
# served at the server root (excluded from the /api/v1 prefix and from this contract).
servers:
- url: http://localhost:3001/api/v1
  description: Local development
- url: https://{host}/api/v1
  description: Deployment
  variables:
    host:
      default: une-platform.example.local
...
paths:
...
  /plans:
    post:
      tags:
      - PLAN
      summary: 계획서 Workspace 생성
      operationId: une_plan_001
      description: '권한: PLAN_CREATE


        핵심 요청: title,startMode,hazardType,managementPhase(templateFileId는 CC-140까지 보류
        — ADR-23 D3)


        핵심 응답: Plan


        오류: PLAN-4001'
      parameters:
      - $ref: '#/components/parameters/CorrelationId'
      - $ref: '#/components/parameters/IdempotencyKeyRequired'
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '503':
          $ref: '#/components/responses/ProviderError'
      x-une-api-id: UNE-PLAN-001
      x-permission: PLAN_CREATE
      x-db-tables:
      - plan
      - document
      x-error-codes:
      - PLAN-4001
      - COM-0400
      - COM-0409
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanCreateRequest'
      security:
      - BearerAuth: []
    get:
      tags:
      - PLAN
      summary: 계획서 목록·검색
      operationId: une_plan_002
      description: '권한: PLAN_READ


        핵심 요청: keyword,status,hazardType,page


        핵심 응답: Page<Plan>


        오류: PLAN-4002'
      parameters:
      - $ref: '#/components/parameters/CorrelationId'
      - name: keyword
        in: query
        required: false
        description: 제목 부분일치 검색어
        schema:
          type: string
          maxLength: 100
      - name: status
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/PlanStatus'
      - name: hazardType
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/PlanHazardType'
      - name: inTrash
        in: query
        required: false
        description: true이면 휴지통(deleted_at NOT NULL) 목록을 조회한다.
        schema:
          type: boolean
          default: false
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
      - name: size
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanPageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '503':
          $ref: '#/components/responses/ProviderError'
      x-une-api-id: UNE-PLAN-002
      x-permission: PLAN_READ
      x-db-tables:
      - plan
      - document
      x-error-codes:
      - PLAN-4002
      security:
      - BearerAuth: []
  /plans/{planId}:
    get:
      tags:
      - PLAN
      summary: 계획서 상세
      operationId: une_plan_003
      description: '권한: PLAN_READ


        핵심 요청: planId


        핵심 응답: PlanDetail


        오류: PLAN-4003'
      parameters:
      - $ref: '#/components/parameters/CorrelationId'
      - name: planId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Success
          headers:
            ETag:
              description: 현재 versionNo (PATCH의 If-Match 값)
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '503':
          $ref: '#/components/responses/ProviderError'
      x-une-api-id: UNE-PLAN-003
      x-permission: PLAN_READ
      x-db-tables:
      - plan
      - document
      x-error-codes:
      - PLAN-4003
      security:
      - BearerAuth: []
    patch:
      tags:
      - PLAN
      summary: 계획서 메타 수정
      operationId: une_plan_004
      description: '권한: PLAN_EDIT


        핵심 요청: If-Match, JSON Merge Patch


        핵심 응답: Plan


        오류: PLAN-409-001


        재시도 안전성은 If-Match 낙관잠금으로 보장(ADR-23 D1 개정)'
      parameters:
      - $ref: '#/components/parameters/CorrelationId'
      - name: planId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          description: Success
          headers:
            ETag:
              description: 갱신된 versionNo
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
        '503':
          $ref: '#/components/responses/ProviderError'
      x-une-api-id: UNE-PLAN-004
      x-permission: PLAN_EDIT
      x-db-tables:
      - plan
      - audit_log
      x-error-codes:
      - PLAN-409-001
      - PLAN-4001
      - PLAN-4003
      - PLAN-412-002
      - COM-0428
      - COM-0400
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanMetaPatchRequest'
      security:
      - BearerAuth: []
    delete:
      tags:
      - PLAN
      summary: 계획서 휴지통 이동
      operationId: une_plan_005
      description: '권한: PLAN_DELETE


        핵심 요청: reason


        핵심 응답: 204


        오류: PLAN-403-001'
      parameters:
      - $ref: '#/components/parameters/CorrelationId'
      - name: planId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: No Content (설계 10 §3.3 응답 204 — 본문 없음)
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '503':
          $ref: '#/components/responses/ProviderError'
      x-une-api-id: UNE-PLAN-005
      x-permission: PLAN_DELETE
      x-db-tables:
      - plan
      x-error-codes:
      - PLAN-403-001
      - PLAN-4003
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanDeleteRequest'
      security:
      - BearerAuth: []
  /plans/{planId}/context-drafts:
    post:
      tags:
      - PLAN
      summary: 기준정보 임시저장
      operationId: une_plan_006
      description: '권한: PLAN_EDIT


        핵심 요청: PlanContextDraft


        핵심 응답: ContextDraft


        오류: PLAN-422-001


        단일 draft upsert(last-write-wins)로 자연 멱등 — 재생 저장소 제외(ADR-23 D1 개정)'
      parameters:
      - $ref: '#/components/parameters/CorrelationId'
      - name: planId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContextDraftResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '503':
          $ref: '#/components/responses/ProviderError'
      x-une-api-id: UNE-PLAN-006
      x-permission: PLAN_EDIT
      x-db-tables:
      - plan_context_draft
      x-error-codes:
      - PLAN-422-001
      - PLAN-4001
      - PLAN-4003
      - PLAN-412-002
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanContextDraftRequest'
      security:
      - BearerAuth: []
  /plans/{planId}/context-snapshots:
    post:
      tags:
      - PLAN
      summary: 기준정보 Snapshot 확정
      operationId: une_plan_007
      description: '권한: PLAN_EDIT


        핵심 요청: PlanContext


        핵심 응답: PlanContextSnapshot


        오류: PLAN-412-001'
      parameters:
      - $ref: '#/components/parameters/CorrelationId'
      - name: planId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/IdempotencyKeyRequired'
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanContextSnapshotResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '503':
          $ref: '#/components/responses/ProviderError'
      x-une-api-id: UNE-PLAN-007
      x-permission: PLAN_EDIT
      x-db-tables:
      - plan_context_snapshot
      - plan
      - audit_log
      x-error-codes:
      - PLAN-412-002
      - PLAN-422-001
      - PLAN-4003
      - COM-0400
      - COM-0409
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanContext'
      security:
      - BearerAuth: []
    get:
      tags:
      - PLAN
      summary: 기준정보 Snapshot 목록
      operationId: une_plan_008
      description: '권한: PLAN_READ


        핵심 요청: -


        핵심 응답: Snapshot[]


        오류: PLAN-404-002'
      parameters:
      - $ref: '#/components/parameters/CorrelationId'
      - name: planId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanContextSnapshotListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '503':
          $ref: '#/components/responses/ProviderError'
      x-une-api-id: UNE-PLAN-008
      x-permission: PLAN_READ
      x-db-tables:
      - plan_context_snapshot
      x-error-codes:
      - PLAN-404-002
      security:
      - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      schema:
        type: string
        maxLength: 100
...
    IdempotencyKeyRequired:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        maxLength: 100
        pattern: '^[A-Za-z0-9._:-]{1,100}$'
    IfMatch:
      name: If-Match
      in: header
      required: false
      schema:
        type: string
  schemas:
...
    Error:
      type: object
      required:
      - code
      - message
      - recoverable
      properties:
        code:
          type: string
        message:
          type: string
        detail:
          type: string
          nullable: true
        recoverable:
          type: boolean
        userAction:
          type: string
          nullable: true
    ErrorEnvelope:
      type: object
      required:
      - success
      - error
      - meta
      properties:
        success:
          type: boolean
          enum:
          - false
        error:
          $ref: '#/components/schemas/Error'
        meta:
          $ref: '#/components/schemas/Meta'
...
    PlanHazardType:
      type: string
      enum:
      - 폭염
      - 태풍/호우
      - 지진
      - 황사
      - 산불
      - 감염병
      - 가축질병
      - 다중밀집건축물붕괴대형사고
      - 정부주요시설
      - 학교시설
    PlanManagementPhase:
      type: string
      enum:
      - 예방
      - 대비
    # 설계 09 §4 계획서 Document 상태모델
    PlanStatus:
      type: string
      enum:
      - DRAFT
      - CONTEXT_READY
      - OUTLINE_GENERATING
      - OUTLINE_REVIEW
      - OUTLINE_CONFIRMED
      - CONTENT_GENERATING
      - EDITING
      - REVIEW_REQUESTED
      - CHANGES_REQUESTED
      - APPROVED
      - FINAL
      - REOPENED
      - ERROR
    PlanStartMode:
      type: string
      enum:
      - BLANK
      - UPLOAD_HWPX
      - RECENT
    PlanResource:
      type: object
      required:
      - planId
      - tenantId
      - title
      - hazardType
      - managementPhase
      - status
      - ownerId
      - versionNo
      - createdAt
      - updatedAt
      properties:
        planId:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 300
        hazardType:
          $ref: '#/components/schemas/PlanHazardType'
        managementPhase:
          $ref: '#/components/schemas/PlanManagementPhase'
        status:
          $ref: '#/components/schemas/PlanStatus'
        documentId:
          type:
          - string
          - 'null'
          format: uuid
        currentContextSnapshotId:
          type:
          - string
          - 'null'
          format: uuid
        currentTocVersionId:
          type:
          - string
          - 'null'
          format: uuid
        startMode:
          $ref: '#/components/schemas/PlanStartMode'
        versionNo:
          type: integer
          description: 낙관적 잠금 버전 (If-Match/ETag 값)
        deletedAt:
          type:
          - string
          - 'null'
          format: date-time
          description: 휴지통 이동 시각 (null이면 활성)
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    # ADR-23 D4: 확정된 기준정보는 불변이며 content_hash(정규화 JSON SHA-256)로
    # 동일 내용 재확정을 dedupe한다.
    PlanContextSnapshotResource:
      type: object
      required:
      - contextSnapshotId
      - planId
      - versionNo
      - contextJson
      - contentHash
      - confirmedBy
      - confirmedAt
      properties:
        contextSnapshotId:
          type: string
          format: uuid
        planId:
          type: string
          format: uuid
        versionNo:
          type: integer
        contextJson:
          $ref: '#/components/schemas/PlanContext'
        contentHash:
          type: string
          pattern: '^[0-9a-f]{64}$'
        supersedesId:
          type:
          - string
          - 'null'
          format: uuid
        confirmedBy:
          type: string
          format: uuid
        confirmedAt:
          type: string
          format: date-time
    # ADR-23 D2: 계획서당 1행 upsert. 임시저장은 완화 검증이므로 contextJson은
    # 미완성 상태를 허용한다(타입·enum·길이 위반은 PLAN-422-001).
...
    PlanResponse:
      type: object
      required:
      - success
      - data
      - meta
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/PlanResource'
        meta:
          type: object
    PlanDetailResponse:
      type: object
      required:
      - success
      - data
      - meta
      properties:
        success:
          type: boolean
        data:
          type: object
          allOf:
          - $ref: '#/components/schemas/PlanResource'
          properties:
            currentContextSnapshot:
              anyOf:
              - $ref: '#/components/schemas/PlanContextSnapshotResource'
              - type: 'null'
              description: 현재 확정 Snapshot (없으면 null)
        meta:
          type: object
    PlanPageResponse:
      type: object
      required:
      - success
      - data
      - meta
      properties:
        success:
          type: boolean
        data:
          type: object
          required:
          - items
          - page
          - size
          - totalElements
          - totalPages
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/PlanResource'
            page:
              type: integer
            size:
              type: integer
            totalElements:
              type: integer
            totalPages:
              type: integer
        meta:
          type: object
    ContextDraftResponse:
      type: object
      required:
      - success
      - data
      - meta
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/ContextDraftResource'
        meta:
          type: object
    PlanContextSnapshotResponse:
      type: object
      required:
      - success
      - data
      - meta
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/PlanContextSnapshotResource'
        meta:
          type: object
    PlanContextSnapshotListResponse:
      type: object
      required:
      - success
      - data
      - meta
      properties:
        success:
          type: boolean
        data:
          type: object
          required:
          - items
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/PlanContextSnapshotResource'
        meta:
          type: object
    # ADR-23 D3: plan.hazard_type/management_phase는 NOT NULL이므로 생성 요청에
    # 필수로 포함한다(기준선 계약 누락 해소).
    PlanCreateRequest:
      type: object
      required:
      - title
      - startMode
      - hazardType
      - managementPhase
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 300
        startMode:
          $ref: '#/components/schemas/PlanStartMode'
        hazardType:
          $ref: '#/components/schemas/PlanHazardType'
        managementPhase:
          $ref: '#/components/schemas/PlanManagementPhase'
        templateFileId:
          type:
          - string
          - 'null'
          format: uuid
          description: 파일 업로드는 CC-140 범위로 보류 — 값이 제공되면 400 PLAN-4001로 거부한다 (ADR-23 D3).
    # JSON Merge Patch 부분집합: 메타 3필드만 허용하고 null(필드 삭제)은 불가.
    # status/owner 변경은 별도 도메인 액션이다 (ADR-23 D5).
    PlanMetaPatchRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 300
        hazardType:
          $ref: '#/components/schemas/PlanHazardType'
        managementPhase:
          $ref: '#/components/schemas/PlanManagementPhase'
    PlanDeleteRequest:
      type: object
      properties:
        reason:
          type: string
          maxLength: 500
    PlanContextDraftRequest:
      type: object
      required:
      - context
      properties:
        context:
          type: object
          additionalProperties: true
          description: 완화 검증 대상 기준정보 작업본 (미완성 허용, ADR-23 D2)
        schemaVersion:
          type: string
          maxLength: 20
          default: '1.0'
    PlanContext:
      type: object
      description: '엄격 검증: contracts/schemas/plan-context.schema.json
        (additionalProperties 불가). Snapshot 확정은 전체 스키마 통과 필수,
        draft는 required/minLength/minItems만 유예'
      additionalProperties: true
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unprocessable:
      description: Unprocessable Entity
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    PreconditionFailed:
      description: Precondition Failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    PreconditionRequired:
      description: Precondition Required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ProviderError:
      description: Provider Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
security:
- BearerAuth: []
...